make: fix sign-compare errors
    cpu, nrf5x_common: fix sign-compare in periph/flashpage
    drivers, periph_common: fix sign-compare in flashpage
    cpu, sam0_common: fix sign-compare error in periph/gpio
    cpu, cc2538: fix sign-compare in periph/timer
    cpu, sam3: fix sign-compare in periph/gpio
    cpu, stm32_common: fix sign-compare in periph/pwm
    cpu, stm32_common: fix sign-compare in periph/timer
    cpu, stm32_common: fix sign-compare in periph/flashpage
    cpu, nrf5x_common: fix sign-compare in radio/nrfmin
    cpu, samd21: fix sign-compare in periph/pwm
    cpu, ezr32wg: fix sign-compare in periph/gpio
    cpu, ezr32wg: fix sign-compare in periph/timer
    drivers, ethos: fix sign-compare
    sys, net: fix sign-compare
    cpu, atmega_common: fix sign-compare error
    cpu, msp430fxyz: fix sign-compare in periph/gpio
    boards, msb-430-common: fix sign-compare in board_init
    driver, cc2420: fix sign-compared
    sys/net: fix sign-compare in gnrc_tftp
    driver, pcd8544: fix sign-compare
    driver, pn532: fix sign-compare
    driver, sdcard_spi: fix sign-compare
    tests: fix sign_compare
    sys/net, lwmac: fix sign_compare
    pkg, lwip: fix sign-compare
    boards, waspmote: make CORECLOCK unsigned long to fix sign_compare error
    tests, sock_ip: fix sign compare
    tests, msg_avail: fix sign compare
    tests, sock_udp: fix sign compare
    boards: fix sign-compare for calliope and microbit matrix

+++ cpu/cc2538/periph/timer.c
@@ -173,7 +173,7 @@ int timer_set_absolute(tim_t tim, int channel, unsigned int value)
 {
     DEBUG("%s(%u, %u, %u)\n", __FUNCTION__, tim, channel, value);
 
-    if ((tim >= TIMER_NUMOF) || (channel >= timer_config[tim].chn) ) {
+    if ((tim >= TIMER_NUMOF) || (channel >= (int)timer_config[tim].chn) ) {
         return -1;
     }
     /* clear any pending match interrupts */
@@ -194,7 +194,7 @@ int timer_clear(tim_t tim, int channel)
 {
     DEBUG("%s(%u, %u)\n", __FUNCTION__, tim, channel);
 
-    if ( (tim >= TIMER_NUMOF) || (channel >= timer_config[tim].chn) ) {
+    if ( (tim >= TIMER_NUMOF) || (channel >= (int)timer_config[tim].chn) ) {
         return -1;
     }
     /* clear interupt flags */
@@ -259,7 +259,7 @@ static void irq_handler(tim_t tim, int channel)
 {
   DEBUG("%s(%u,%d)\n", __FUNCTION__, tim, channel);
   assert(tim < TIMER_NUMOF);
-  assert(channel < timer_config[tim].chn);
+  assert(channel < (int)timer_config[tim].chn);
 
   uint32_t mis;
   /* Latch the active interrupt flags */
