+++ cpu/stm32_common/periph/flashpage.c
@@ -29,7 +29,7 @@
 
 void flashpage_write(int page, void *data)
 {
-    assert(page < FLASHPAGE_NUMOF);
+    assert(page < (int)FLASHPAGE_NUMOF);
 
     uint16_t *page_addr = flashpage_addr(page);
     uint16_t *data_addr = (uint16_t *)data;
