+++ drivers/periph_common/flashpage.c
@@ -30,14 +30,14 @@
 
 void flashpage_read(int page, void *data)
 {
-    assert(page < FLASHPAGE_NUMOF);
+    assert(page < (int)FLASHPAGE_NUMOF);
 
     memcpy(data, flashpage_addr(page), FLASHPAGE_SIZE);
 }
 
 int flashpage_verify(int page, void *data)
 {
-    assert(page < FLASHPAGE_NUMOF);
+    assert(page < (int)FLASHPAGE_NUMOF);
 
     if (memcmp(flashpage_addr(page), data, FLASHPAGE_SIZE) == 0) {
         return FLASHPAGE_OK;
